+++ app/src/main/java/com/example/toyproject_dpm/phonebook/UserDTO.kt
@@ -1,0 +1,9 @@
+package com.example.toyproject_dpm.phonebook
+
+data class UserDTO(
+    var Name: String,
+    var PhoneNum: String,
+    var age: Int,
+    var Country: String
+
+)
